+++ src/Calc-ChemWeightRL.py
@@ -31,7 +31,7 @@ def raise_error(msg, *arg):
     sys.exit(1)
 
 def get_args():
-    help_desc_msg ="""Calc-ChemWeightRL - Chemical weighing calculator using pymatgen
+    help_desc_msg = r"""Calc-ChemWeightRL - Chemical weighing calculator using pymatgen
 
     Synopsis:
         Calculate the weight(gram) of each term in the chemical formula
@@ -74,7 +74,7 @@ def get_args():
         Calc-ChemMassPercent.py
 
     """
-    help_epi_msg = """EXAMPLES:
+    help_epi_msg = r"""EXAMPLES:
 
     # Calculate Grams of Methane(CH4) and Oxygen(O2) required to obtain 44g of carbon dioxide(CO2)
     "CH4 + 2 O2 -> CO2 + 2 H2O" | python Calc-ChemWeightRL.py -g 44.0